Cleanup handler uses only defined variables
As there are many exit points, Cleanup can be called
when some variables are not set yet. Instead of moving
all vars init to the start, Cleanup can check for variable
existence, this also avoids running some cleanup commands.

+++ otc-tools-tester-run-test-kvm.sh
@@ -4,10 +4,12 @@ UMOUNT="sudo umount -l"
 OBS_DELETION="$WORKSPACE/info"
 
 Cleanup () {
- mountpoint -q $BUILDMOUNT && $UMOUNT $BUILDMOUNT
- mountpoint -q $KVM_ROOT && $UMOUNT $KVM_ROOT
- rm -fr $KVM_ROOT
- rm -fr $SRC_TMPCOPY
+ test "${SRC_TMPCOPY+defined}" && rm -fr $SRC_TMPCOPY
+ test "${BUILDMOUNT+defined}" && mountpoint -q $BUILDMOUNT && $UMOUNT $BUILDMOUNT
+ if test "${KVM_ROOT+defined}" ; then
+     mountpoint -q $KVM_ROOT && $UMOUNT $KVM_ROOT
+     rm -fr $KVM_ROOT
+ fi
  date
 }
